refactor(#1390): created DelayedAtomicConstraint interface
