+++ common/src/main/java/com/scottlogic/deg/common/profile/constraints/delayed/DelayedAtomicConstraint.java
@@ -14,63 +14,16 @@
  * limitations under the License.
  */
 
-
 package com.scottlogic.deg.common.profile.constraints.delayed;
 
-import com.scottlogic.deg.common.date.TemporalAdjusterGenerator;
 import com.scottlogic.deg.common.profile.Field;
 import com.scottlogic.deg.common.profile.constraintdetail.AtomicConstraintType;
 import com.scottlogic.deg.common.profile.constraints.Constraint;
 
-public class DelayedAtomicConstraint implements Constraint {
+public interface DelayedAtomicConstraint extends Constraint {
+    Field getField();
 
-    private final Field field;
-    private final AtomicConstraintType underlyingConstraint;
-    private final Field otherField;
+    AtomicConstraintType getUnderlyingConstraint();
 
-    private final TemporalAdjusterGenerator offsetGenerator;
-
-    private final Integer offsetUnit;
-    public DelayedAtomicConstraint(Field field, AtomicConstraintType underlyingConstraint, Field otherField, TemporalAdjusterGenerator offsetGenerator, Integer offsetUnit) {
-        this.field = field;
-        this.underlyingConstraint = underlyingConstraint;
-        this.otherField = otherField;
-        this.offsetGenerator = offsetGenerator;
-        this.offsetUnit = offsetUnit;
-    }
-
-    public DelayedAtomicConstraint(Field field, AtomicConstraintType underlyingConstraint, Field otherField) {
-        this(field, underlyingConstraint, otherField, null, null);
-    }
-
-    static void validateFieldsAreDifferent(Field first, Field second) {
-        if (first.equals(second)) {
-            throw new IllegalArgumentException("Cannot have a relational field referring to itself");
-        }
-    }
-
-    public Field getField(){
-        return field;
-    }
-
-    public AtomicConstraintType getUnderlyingConstraint(){
-        return underlyingConstraint;
-    }
-
-    public Field getOtherField(){
-        return otherField;
-    }
-
-    public DelayedAtomicConstraint negate() {
-        return new DynamicNotConstraint(this);
-    }
-
-    public TemporalAdjusterGenerator getOffsetGenerator() {
-        return offsetGenerator;
-    }
-
-    public Integer getOffsetUnit() {
-        return offsetUnit;
-    }
-
+    Field getOtherField();
 }
